requirements.txt
Django>=1.11,<2.0
django-simple-history
httplib2
apiclient
google-api-python-client
xlrd
xlutils
psycopg2-binary
celery
redis
gunicorn
django-prettyjson
djangorestframework
